+++ src/components/interfaces/noora-chat/chat/Messages.tsx
@@ -6,7 +6,7 @@ import { InformationCircleIcon } from "@heroicons/react/outline";
 
 import dynamic from "next/dynamic";
 import { ACTIONS, EVENTS, STATUS } from "react-joyride";
-import SpeechSynthesizer from "../../../global/utility/SpeechSynthesizer";
+import SpeechSynthesizer from "../../speech/SpeechSynthesizer";
 import { messageToSpeechParams } from "../../../../scripts/util";
 const JoyRideNoSSR = dynamic(() => import("react-joyride"), { ssr: false });
