+++ src/components/interfaces/speech/SpeechSynthesizer.tsx
